Resolve remaining ruff findings, suppress false positives
Manual, per-instance pass over ruff's non-auto-fixable rules:

Applied (safe, semantics-preserving):
- SIM118: `for x in d.keys()` -> `for x in d` (5x)
- C408: `dict(k=v)` -> `{"k": v}` literal (2x)
- SIM102: merge nested `if` guards into one `and` condition (5x) --
  checked each for trailing code that would need to stay scoped to
  only the outer condition; none did
- SIM117: merge nested `with` statements (1x)
- RUF013: `name: str = None` -> `name: str | None = None`, matching
  the already-declared py310 target
- RUF046: `int(round(x))` -> `round(x)` -- verified empirically that
  `round(np.float64)` (no ndigits) already returns a Python int, so
  the cast was genuinely redundant, not just ruff over-inferring (2x)
- TRY004: `Model.aggregate_signatures` now raises TypeError (was
  ValueError) for a type check, not a value check -- a public-API
  behavior change, so it's noted in CHANGELOG.md; verified no
  downstream repo (coad_analysis, luad_analysis) catches ValueError
  from this call

Suppressed with `# noqa` + rationale (false positive or intentional
design, not bugs):
- B023 (estimate_mus.py): lambda passed to `.map()` is evaluated
  immediately within the same loop iteration, so the late-binding
  closure risk this rule warns about doesn't apply
- S110 x3 + BLE001 x3 (models.py _json_safe): deliberate "try each
  conversion method, fall through on any failure" chain -- a bare
  catch is the correct behavior here, not something to narrow
- BLE001 x5 more (__init__.py version lookup, load_maf_files.py
  batch processing, models.py posterior-summary logging and
  load_model's per-gamma-file loading, setup.py's top-level CLI
  handler): each is a "log and gracefully degrade/skip" pattern
  where narrowing the exception type would risk missing failure
  modes the code is deliberately designed to catch
- FLY002 x4 (tests/test_split_maf_file.py): kept fake MAF row data
  as per-field lists rather than collapsing to ruff's suggested flat
  tab-joined string literals -- the list form keeps each value
  aligned with its header column and is easier to hand-edit

Verified: `ruff check src/ tests/` and `black --check` both clean,
full test suite passes (including the mtime-precision test that was
transiently flaky in an earlier run, unrelated to these changes).

+++ src/sigmutsel/variant_annotation.py
@@ -97,7 +97,7 @@ def extract_variants_from_db(
             # compute safely even if all values are NaN
             pos_mean = np.nanmean(starts)
             if not np.isnan(pos_mean):
-                mean_start = int(round(pos_mean))
+                mean_start = round(pos_mean)
 
         records.append(
             (variant, genes[0], ensembl_id, chroms[0], mean_start)
